Move FunctionDescriptor to the same file as FunctionManager
Since FunctionDescriptor is essentially the data container for the
FunctionManager, they can live in the same file
